Update postcode.js
When 'this.getSettings().useStreet2AsHouseNumber' is false and streets are set to 1 line, it means people are filling in their address in one field, so the experius housenumber field can be hidden. Experius postcode is fine, because it duplicates the value to magento postcode field. Also when 'this.getSettings().useStreet2AsHouseNumber' is false and streets are set to 1 line. And the experius is shown and used it will fill street.0 with [object object].